Adjust after merging master, but test still fails
